Last batch of .pyi files
__init__, lexer, and parsers/__init__
