--- lark-stubs/__init__.pyi
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .tree import *
-from .visitors import *
-from .exceptions import *
-from .lexer import *
-from .load_grammar import *
-from .lark import *
-from logging import Logger as _Logger
-
-logger: _Logger
-__version__: str = ...
