+++ lark/__init__.py
@@ -6,4 +6,4 @@ from .exceptions import (ParseError, LexError, GrammarError, UnexpectedToken,
 from .lexer import Token
 from .lark import Lark
 
-__version__ = "1.0.0a"
+__version__: str = "1.0.0a"
